daemon: extract tail reconcile phase from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
tail-reconcile orchestrator and its LLDP, DHCP-relay, event-engine, and
initial policy-scheduler reconcile helpers move verbatim into
daemon_apply_tail.go. applyTailReconciles drives these helpers, so they
form one cohesive tail-phase unit.

Moved declarations (byte-identical bodies, no rename, no logic change):
applyTailReconciles, reconcileDHCPRelay, effectiveLLDPConfig,
reconcileLLDP, initEventEngine, reconcileEventOptions, lldpConfigEqual,
publishInitialPolicySchedulerStateLocked. Same package, no caller
changes; per-file imports pruned by goimports.

After this final increment daemon_apply.go retains only the apply
entrypoints (applyConfig/applyConfigResult/applyCancelCtx), the core
applyConfigLocked orchestrator, the two procfs knob helpers, and
compileErrorMustAbortApply. Build: go build ./pkg/daemon/... green.
